ajustes en bash profile y emacs para macos

+++ .bash_profile
@@ -6,8 +6,8 @@ if [ -f ${HOME}/.bashrc ]; then
 fi
 
 # ssh-agent
-if [ -f ${HOME}/.ssh/id_rsa -a -z "$DESKTOP_SESSION" ]; then
-    eval $(/usr/bin/ssh-agent -s) > /dev/null
-    trap "kill ${SSH_AGENT_PID}" 0
-    /usr/bin/ssh-add ${HOME}/.ssh/id_rsa
+if [ -f ${HOME}/.ssh/id_rsa -a -z "$DESKTOP_SESSION" -a $(uname) != "Darwin" ]; then
+   eval $(/usr/bin/ssh-agent -s) > /dev/null
+   trap "kill ${SSH_AGENT_PID}" 0
+   /usr/bin/ssh-add ${HOME}/.ssh/id_rsa
 fi
